+++ px-checkout/src/main/java/com/mercadopago/android/px/internal/extensions/BaseExtensions.kt
@@ -1,0 +1,7 @@
+package com.mercadopago.android.px.internal.extensions
+
+import android.text.TextUtils
+
+fun CharSequence?.isNotNullNorEmpty() = !isNullOrEmpty()
+
+fun CharSequence?.orIfEmpty(fallback: String) = if (isNotNullNorEmpty()) this!! else fallback