+++ px-checkout/src/main/java/com/mercadopago/android/px/internal/features/express/ExpressPayment.java
@@ -1,18 +1,16 @@
 package com.mercadopago.android.px.internal.features.express;
 
 import android.support.annotation.NonNull;
-import com.mercadopago.android.px.addons.model.SecurityValidationData;
 import com.mercadopago.android.px.core.DynamicDialogCreator;
 import com.mercadopago.android.px.internal.base.MvpView;
-import com.mercadopago.android.px.internal.callbacks.PaymentServiceHandler;
-import com.mercadopago.android.px.internal.features.explode.ExplodeDecorator;
 import com.mercadopago.android.px.internal.features.express.installments.InstallmentRowHolder;
 import com.mercadopago.android.px.internal.features.express.slider.HubAdapter;
+import com.mercadopago.android.px.internal.features.pay_button.PayButton;
 import com.mercadopago.android.px.internal.view.ElementDescriptorView;
-import com.mercadopago.android.px.internal.viewmodel.PayButtonViewModel;
+import com.mercadopago.android.px.internal.viewmodel.BusinessPaymentModel;
+import com.mercadopago.android.px.internal.viewmodel.PaymentModel;
 import com.mercadopago.android.px.internal.viewmodel.SplitSelectionState;
 import com.mercadopago.android.px.internal.viewmodel.drawables.DrawableFragmentItem;
-import com.mercadopago.android.px.model.Card;
 import com.mercadopago.android.px.model.Currency;
 import com.mercadopago.android.px.model.DiscountConfigurationModel;
 import com.mercadopago.android.px.model.IPaymentDescriptor;
@@ -23,7 +21,7 @@ import com.mercadopago.android.px.model.Site;
 import com.mercadopago.android.px.model.StatusMetadata;
 import com.mercadopago.android.px.model.exceptions.MercadoPagoError;
 import com.mercadopago.android.px.model.internal.DisabledPaymentMethod;
-import com.mercadopago.android.px.tracking.internal.model.Reason;
+import com.mercadopago.android.px.model.internal.PaymentConfiguration;
 import java.util.List;
 
 public interface ExpressPayment {
@@ -40,36 +38,7 @@ public interface ExpressPayment {
 
         void cancel();
 
-        void showSecurityCodeScreen(@NonNull final Card card,
-            final Reason reason);
-
         void handlePaymentRecovery(@NonNull PaymentRecovery paymentRecovery);
-
-        void showSecurityCodeScreenForRecovery(@NonNull PaymentRecovery paymentRecovery);
-
-        void showPaymentProcessor();
-
-        void finishLoading(@NonNull final ExplodeDecorator params);
-
-        void cancelLoading();
-
-        void startLoadingButton(final int paymentTimeout, @NonNull final PayButtonViewModel payButtonViewModel);
-
-        //TODO shared with Checkout activity
-
-        void showErrorScreen(@NonNull final MercadoPagoError error);
-
-        void showPaymentResult(@NonNull final IPaymentDescriptor paymentResult);
-
-        void startSecurityValidation(@NonNull SecurityValidationData data);
-
-        void startPayment();
-
-        void enableToolbarBack();
-
-        void disableToolbarBack();
-
-        void showErrorSnackBar(@NonNull final MercadoPagoError error);
 
         void updateViewForPosition(final int paymentMethodIndex,
             final int payerCostSelected,
@@ -87,39 +56,27 @@ public interface ExpressPayment {
         void showDisabledPaymentMethodDetailDialog(@NonNull final DisabledPaymentMethod disabledPaymentMethod,
             @NonNull final StatusMetadata currentStatus);
 
-        boolean isExploding();
-
         void resetPagerIndex();
 
         void showDynamicDialog(@NonNull final DynamicDialogCreator creatorFor,
             @NonNull final DynamicDialogCreator.CheckoutData checkoutData);
 
-        void setPayButtonText(@NonNull final PayButtonViewModel payButtonViewModel);
-
         void showOfflineMethods(@NonNull final OfflinePaymentTypesMetadata offlineMethods);
 
         void updateBottomSheetStatus(final boolean hasToExpand);
+
+        void showPaymentResult(@NonNull final PaymentModel model, @NonNull final PaymentConfiguration paymentConfiguration);
+
+        void showBusinessResult(@NonNull final BusinessPaymentModel model);
     }
 
-    interface Actions extends PaymentServiceHandler {
+    interface Actions {
 
         void trackExpressView();
 
-        void startSecuredPayment();
-
-        void confirmPayment();
-
-        void trackSecurityFriction();
-
         void cancel();
 
-        void onTokenResolved();
-
         void loadViewModel();
-
-        void onViewResumed();
-
-        void onViewPaused();
 
         void onInstallmentsRowPressed();
 
@@ -129,10 +86,6 @@ public interface ExpressPayment {
 
         void onPayerCostSelected(final PayerCost payerCostSelected);
 
-        void hasFinishPaymentAnimation();
-
-        void manageNoConnection();
-
         void onSplitChanged(boolean isChecked);
 
         void onHeaderClicked();
@@ -141,6 +94,10 @@ public interface ExpressPayment {
 
         void onOtherPaymentMethodClickableStateChanged(boolean state);
 
-        void handlePaymentRecovery(@NonNull final PaymentRecovery paymentRecovery);
+        void onPaymentProcessingError(@NonNull final MercadoPagoError error);
+
+        void onPaymentFinished(@NonNull final IPaymentDescriptor payment);
+
+        void requireCurrentConfiguration(@NonNull final PayButton.OnReadyForPaymentCallback callback);
     }
 }