+++ px-checkout/src/main/java/com/mercadopago/android/px/internal/features/payment_result/remedies/RemedyState.kt
@@ -1,0 +1,10 @@
+package com.mercadopago.android.px.internal.features.payment_result.remedies
+
+import com.mercadopago.android.px.internal.features.payment_result.remedies.view.CvvRemedy
+import com.mercadopago.android.px.internal.viewmodel.PaymentModel
+
+internal sealed class RemedyState {
+    internal data class ShowCvvRemedy(val model: CvvRemedy.Model): RemedyState()
+    internal object ShowKyCRemedy: RemedyState()
+    internal data class ShowResult(val paymentModel: PaymentModel): RemedyState()
+}