+++ px-checkout/src/main/java/com/mercadopago/android/px/internal/repository/CongratsRepository.kt
@@ -1,0 +1,14 @@
+package com.mercadopago.android.px.internal.repository
+
+import com.mercadopago.android.px.internal.viewmodel.PaymentModel
+import com.mercadopago.android.px.model.IPaymentDescriptor
+import com.mercadopago.android.px.model.PaymentResult
+
+interface CongratsRepository {
+    fun getPostPaymentData(payment: IPaymentDescriptor, paymentResult: PaymentResult,
+                           callback: PostPaymentCallback)
+
+    interface PostPaymentCallback {
+        fun handleResult(paymentModel: PaymentModel)
+    }
+}