+++ px-checkout/src/main/java/com/mercadopago/android/px/tracking/internal/model/ConfirmData.java
@@ -1,5 +1,6 @@
 package com.mercadopago.android.px.tracking.internal.model;
 
+import android.os.Parcel;
 import android.support.annotation.Keep;
 import android.support.annotation.NonNull;
 import com.mercadopago.android.px.tracking.internal.events.ConfirmEvent;
@@ -10,6 +11,18 @@ public class ConfirmData extends AvailableMethod {
 
     private final String reviewType;
     private int paymentMethodSelectedIndex;
+
+    public static final Creator<ConfirmData> CREATOR = new Creator<ConfirmData>() {
+        @Override
+        public ConfirmData createFromParcel(final Parcel in) {
+            return new ConfirmData(in);
+        }
+
+        @Override
+        public ConfirmData[] newArray(final int size) {
+            return new ConfirmData[size];
+        }
+    };
 
     public ConfirmData(@NonNull final ConfirmEvent.ReviewType reviewType, final int paymentMethodSelectedIndex,
         @NonNull final AvailableMethod availableMethod) {
@@ -22,5 +35,24 @@ public class ConfirmData extends AvailableMethod {
         @NonNull final AvailableMethod availableMethod) {
         super(availableMethod.paymentMethodId, availableMethod.paymentMethodType, availableMethod.extraInfo);
         this.reviewType = reviewType.value;
+    }
+
+    @SuppressWarnings("WeakerAccess")
+    protected ConfirmData(final Parcel in) {
+        super(in);
+        reviewType = in.readString();
+        paymentMethodSelectedIndex = in.readInt();
+    }
+
+    @Override
+    public void writeToParcel(final Parcel dest, final int flags) {
+        super.writeToParcel(dest, flags);
+        dest.writeString(reviewType);
+        dest.writeInt(paymentMethodSelectedIndex);
+    }
+
+    @Override
+    public int describeContents() {
+        return 0;
     }
 }