+++ px-checkout/src/test/java/com/mercadopago/android/px/internal/features/pay_button/PayButtonViewModelTest.kt
@@ -1,0 +1,86 @@
+package com.mercadopago.android.px.internal.features.pay_button
+
+import org.junit.Before
+import org.junit.Test
+import org.junit.runner.RunWith
+import org.mockito.junit.MockitoJUnitRunner
+
+@RunWith(MockitoJUnitRunner::class)
+class PayButtonViewModelTest {
+
+    @Before
+    fun setUp() {
+    }
+
+    @Test
+    fun getButtonTextLiveData() {
+    }
+
+    @Test
+    fun getPaymentStartedLiveData() {
+    }
+
+    @Test
+    fun getCvvRequiredLiveData() {
+    }
+
+    @Test
+    fun getRecoverRequiredLiveData() {
+    }
+
+    @Test
+    fun getStateUILiveData() {
+    }
+
+    @Test
+    fun attach() {
+    }
+
+    @Test
+    fun startSecuredPayment() {
+    }
+
+    @Test
+    fun handleBiometricsResult() {
+    }
+
+    @Test
+    fun startPayment() {
+    }
+
+    @Test
+    fun onPaymentError() {
+    }
+
+    @Test
+    fun onVisualPayment() {
+    }
+
+    @Test
+    fun onCvvRequired() {
+    }
+
+    @Test
+    fun onPaymentFinished() {
+    }
+
+    @Test
+    fun onRecoverPaymentEscInvalid() {
+    }
+
+    @Test
+    fun recoverPayment() {
+    }
+
+    @Test
+    fun testRecoverPayment() {
+    }
+
+    @Test
+    fun preparePayment() {
+    }
+
+    @Test
+    fun hasFinishPaymentAnimation() {
+    }
+}